+++ quint/src/idGenerator.ts
@@ -22,9 +22,9 @@ export interface IdGenerator {
 /**
  * Introduce a new id generator.
  */
-export const newIdGenerator = () => {
+export const newIdGenerator = (initialId: bigint = 0n) => {
   return {
-    _lastId: 0n,
+    _lastId: initialId,
 
     nextId(): bigint {
         this._lastId = this._lastId + 1n
@@ -32,4 +32,3 @@ export const newIdGenerator = () => {
     },
   }
 }
-